# perf(css/minifier): Use `JsWord` instead of `&str`