--- a/crates/swc_atoms/scripts/add-from-cargo-check.sh
+++ b/crates/swc_atoms/scripts/add-from-cargo-check.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -eu
+
+export CARGO_TERM_COLOR=never
+
+cargo check 2>&1 | grep 'no rules expected the token' | sed -e 's/error: no rules expected the token `"//' | sed -e 's/"`//'