--- a/crates/swc_atoms/words.txt
+++ b/crates/swc_atoms/words.txt
@@ -675,3 +675,7 @@ void
 while
 with
 yield
+opacity
+fill-opacity
+stroke-opacity
+shape-image-threshold