--- a/crates/swc_css_minifier/src/compress/color.rs
+++ b/crates/swc_css_minifier/src/compress/color.rs
@@ -378,7 +378,7 @@ impl CompressColor {
                 Some((2.55 * *value).round())
             }
             Some(ComponentValue::Ident(Ident { value, .. }))
-                if &*value.to_ascii_lowercase() == js_word!("none") =>
+                if value.to_ascii_lowercase() == js_word!("none") =>
             {
                 Some(0.0)
             }