--- a/crates/swc_css_minifier/src/compress/easing_function.rs
+++ b/crates/swc_css_minifier/src/compress/easing_function.rs
@@ -1,3 +1,4 @@
+use swc_atoms::js_word;
 use swc_css_ast::*;
 use swc_css_visit::{VisitMut, VisitMutWith};
 
@@ -18,7 +19,9 @@ impl VisitMut for CompressEasingFunction {
                 name,
                 value: function_value,
                 span,
-            }) if &*name.value.to_lowercase() == "cubic-bezier" && function_value.len() == 7 => {
+            }) if name.value.to_ascii_lowercase() == js_word!("cubic-bezier")
+                && function_value.len() == 7 =>
+            {
                 if let (
                     ComponentValue::Number(Number { value: first, .. }),
                     ComponentValue::Number(Number { value: second, .. }),
@@ -67,7 +70,9 @@ impl VisitMut for CompressEasingFunction {
                 name,
                 value: function_value,
                 span,
-            }) if &*name.value.to_lowercase() == "steps" && function_value.len() == 3 => {
+            }) if name.value.to_ascii_lowercase() == js_word!("steps")
+                && function_value.len() == 3 =>
+            {
                 match (&function_value[0], &function_value[2]) {
                     (
                         ComponentValue::Number(Number {