**Description:**

```
Benchmarking es/minify/libraries/bootstrap
Benchmarking es/minify/libraries/bootstrap: Warming up for 3.0000 s
Benchmarking es/minify/libraries/bootstrap: Collecting 10 samples in estimated 5.0937 s (440 iterations)
Benchmarking es/minify/libraries/bootstrap: Analyzing
es/minify/libraries/bootstrap
                        time:   [11.497 ms 11.533 ms 11.564 ms]
                        change: [-5.7403% -5.1161% -4.5238%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```